split BEWGor.py script into tractable small scripts
