+++ bewgor/banners.py
@@ -1,0 +1,73 @@
+#!/usr/bin/env python3
+
+# Function to show the big ASCII Art
+def showTheName():
+	print ("""
+  ___          _   _   _          ___
+ | _ )  _  _  | | | | ( )  ___   | __|  _  _   ___
+ | _ \ | || | | | | | |/  (_-<   | _|  | || | / -_)
+ |___/  \_,_| |_| |_|     /__/   |___|  \_, | \___|
+ __      __                 _   _   _   |__/ _
+ \ \    / /  ___   _ _   __| | | | (_)  ___ | |_
+  \ \/\/ /  / _ \ | '_| / _` | | | | | (_-< |  _|
+   \_/\_/   \___/ |_|   \__,_| |_| |_| /__/  \__|
+  / __|  ___   _ _    ___   _ _   __ _  | |_   ___   _ _
+ | (_ | / -_) | ' \  / -_) | '_| / _` | |  _| / _ \ | '_|
+  \___| \___| |_||_| \___| |_|   \__,_|  \__| \___/ |_|
+ also knowns as BEWGor
+ Pronounced "Booger"
+ "Pick your Knows!"
+
+
+ by berzerk0 @ Github
+ Version 0.0.1 - 'FLAT' ALPHA RELEASE
+        """)
+
+#Function to show the mini ASCII Art
+def showMiniName():
+	print ("""
+  ___ _____      _____
+ | _ ) __\ \    / / __|___ _ _
+ | _ \ _| \ \/\/ / (_ / _ \ '_|
+ |___/___| \_/\_/ \___\___/_| """ )
+
+
+#shows the BEWGor logo in ASCII art
+def showNose():
+	print ("""
+                          .`
+
+        :
+                         -.
+                        `+
+                        +.
+                       -+
+                      `s
+                     `s.
+                     s:
+                   `s/
+                  `s/
+                 -y:
+               `+y.
+              :y+`
+            -ss.
+          -ss-
+        .sy:                         ```
+      `+h/`                          ```.-.`
+     -hs.                                 .-:`
+    :d/                                     `/-
+   /m:                                        o-
+  .m+                                         .s
+  sd                                          .y
+  ds                                          s:
+  dy                                        `oo
+  +m.                   ``.::::-.``       `/s:
+   yd.              `-+shddhhhhhsooso+//+oo:`
+    +do.`      ``:oyyo:yys-..-oys` -hy-.`
+     .+hyo+++osyyohh. -yy-`yh .yy: `yy
+        `-:::-.   +y/ `syoy:.-oys` -ys
+                  `sy: `yyyyyys/` -yy.
+                   `oy+y- `..`  .+yo.
+                    `ssyyo+//+osy+-
+                  `sd+  .-://:-.
+                   `-              """)
